初步完成应用页面
