+++ app.json
@@ -1,8 +1,8 @@
 {
   "pages": [
+    "pages/application/application",
     "pages/index/index",
     "pages/news/news",
-    "pages/application/application",
     "pages/person/person",
     "pages/logs/logs"
   ],
