+++ pages/application/application.wxss
@@ -1,16 +1,34 @@
 /* pages/application/application.wxss */
-.iconbox{
+
+.bar-icon {
   display: flex;
-  flex-direction: column;
+  flex-direction: row;
+  box-sizing: border-box;
+  align-items: center;
+  width: 750rpx;
   height: 60rpx;
-  width: 40rpx;
+  background-color: #F5F5F5;
 }
-.myicon{
+
+.bar-name {
+  display: flex;
+  flex-direction: row;
+  box-sizing: border-box;
+  align-items: center;
+  width: 750rpx;
+  height: 60rpx;
+  background-color: #F5F5F5;
+}
+
+.myicon {
   height: 50rpx;
   width: 50rpx;
+  margin: 0 auto;
 }
-.mylabel{
-  font-size: 20rpx;
+
+.myname {
+  font-size: 30rpx;
+  margin: 0 auto;
 }
 
 .boxMode0 {
@@ -31,7 +49,7 @@
 
 .boxMode1 {
   width: 750rpx;
-  height: 475rpx;
+  height: 550rpx;
   border-top: 5rpx solid #A9A9A9;
 }
 
@@ -44,6 +62,7 @@
   padding-left: 5rpx;
   margin-left: 20rpx;
   margin-top: 10rpx;
+  margin-bottom: 10rpx;
   border-radius: 10rpx;
 }
 
@@ -54,13 +73,13 @@
 
 .boxMode2 {
   width: 750rpx;
-  height: 300rpx;
+  height: 320rpx;
   border-top: 5rpx solid #A9A9A9;
 }
 
 .boxMode3 {
   width: 750rpx;
-  height: 200rpx;
+  height: 250rpx;
   border-top: 5rpx solid #A9A9A9;
   border-bottom: 5rpx solid #A9A9A9;
 }
