+++ pages/index/index.wxss
@@ -1,30 +1,2 @@
 /**index.wxss**/
-.boxMode1 {
-  display: flex;
-  flex-direction: row;
-  width: 550rpx;
-  height: 80rpx;
-  background-color: #F5F5F5;
-  padding: 20rpx 40rpx;
-  /*内边距*/
-  margin: 400rpx 60rpx;
-  /*外边距*/
-  border-radius: 30rpx;
-    /*圆角边框*/
-}
 
-.myleft {
-  height: 80rpx;
-  width:80rpx;
-}
-
-.mycenter {
-  font-size: 60rpx;
-  margin-left: 30rpx;
-}
-
-.myright {
-  height: 80rpx;
-  width:80rpx;
-  margin-left:240rpx;
-}
