+++ pages/logs/logs.wxml
@@ -1,6 +1,2 @@
-<!--logs.wxml-->
-<view class="container log-list">
-  <block wx:for="{{logs}}" wx:for-item="log">
-    <text class="log-item">{{index + 1}}. {{log}}</text>
-  </block>
-</view>
+<!--pages/logs/logs.wxml-->
+<text>pages/logs/logs.wxml</text>
