+++ pages/logs/logs.wxss
@@ -1,8 +1,1 @@
-.log-list {
-  display: flex;
-  flex-direction: column;
-  padding: 40rpx;
-}
-.log-item {
-  margin: 10rpx;
-}
+/* pages/logs/logs.wxss */
