+++ pages/person/person.js
@@ -16,33 +16,14 @@ Page({
         userInfo: app.globalData.userInfo,
         hasUserInfo: true
       })
-    } else if (this.data.canIUse) {
-      // 由于 getUserInfo 是网络请求，可能会在 Page.onLoad 之后才返回
-      // 所以此处加入 callback 以防止这种情况
-      app.userInfoReadyCallback = res => {
-        this.setData({
-          userInfo: res.userInfo,
-          hasUserInfo: true
-        })
-      }
-    } else {
-      // 在没有 open-type=getUserInfo 版本的兼容处理
-      wx.getUserInfo({
-        success: res => {
-          app.globalData.userInfo = res.userInfo
-          this.setData({
-            userInfo: res.userInfo,
-            hasUserInfo: true
-          })
-        }
-      })
     }
   },
-  getUserInfo(e) {
-    console.log(e)
-    app.globalData.userInfo = e.detail.userInfo
+
+  // 通过参数 userInfo 得到用户信息，被 `./person.wxml` 中登录按钮使用。
+  getUserInfo(userInfo) {
+    app.globalData.userInfo = userInfo.detail.userInfo
     this.setData({
-      userInfo: e.detail.userInfo,
+      userInfo: userInfo.detail.userInfo,
       hasUserInfo: true
     })
   }
